# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LifeFlow API
  description: |
    Personal productivity API for task management, goal tracking, habits,
    focus sessions, lists, notes, and more.

    ## Error Responses
    All error responses use a consistent shape:
    ```json
    { "error": "Human-readable error message" }
    ```

    ## Common Status Codes
    | Code | Meaning |
    |------|---------|
    | 200  | Success |
    | 201  | Created (POST that creates a resource) |
    | 400  | Bad request (validation error) |
    | 404  | Resource not found |
    | 409  | Conflict (constraint violation) |
    | 500  | Internal server error |
  version: 0.7.1
  license:
    name: MIT
servers:
  - url: http://localhost:3456
    description: Local development
...
paths:
...
  /api/notes:
    get:
      tags: [Notes]
      summary: List all notes
      responses:
        '200':
          description: Array of notes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Note' }
    post:
      tags: [Notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                content: { type: string }
                pinned: { type: integer, enum: [0, 1] }
                color: { type: string }
      responses:
        '201':
          description: Created note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
  /api/notes/{id}:
    get:
      tags: [Notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Note object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
    put:
      tags: [Notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
    delete:
      tags: [Notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }

  # ─── Inbox ───
  /api/inbox:
    get:
      tags: [Inbox]
      summary: List inbox items
      responses:
        '200':
          description: Array of inbox items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InboxItem' }
    post:
      tags: [Inbox]
      summary: Quick-capture to inbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '201':
          description: Created inbox item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboxItem' }
  /api/inbox/{id}:
    put:
      tags: [Inbox]
      summary: Update an inbox item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated inbox item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboxItem' }
    delete:
      tags: [Inbox]
      summary: Delete an inbox item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/inbox/{id}/triage:
    post:
      tags: [Inbox]
      summary: Triage inbox item to a task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goal_id]
              properties:
                goal_id: { type: integer }
                priority: { type: integer }
                due_date: { type: string, format: date }
      responses:
        '200':
          description: Created task from inbox item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }

  # ─── Reviews ───
  /api/reviews:
    get:
      tags: [Reviews]
      summary: List weekly reviews
      responses:
        '200':
          description: Array of reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }
    post:
      tags: [Reviews]
      summary: Create a weekly review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wins: { type: string }
                challenges: { type: string }
                next_week: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
      responses:
        '201':
          description: Created review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
...
  /api/reviews/{id}:
    delete:
      tags: [Reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/reviews/daily:
    post:
      tags: [Reviews]
      summary: Create or update daily micro-review
      description: Upserts a daily review entry for the given date.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date]
              properties:
                date: { type: string, format: date }
                note: { type: string }
                completed_count: { type: integer }
      responses:
        '200':
          description: Daily review created or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  date: { type: string, format: date }
                  note: { type: string }
                  completed_count: { type: integer }
  /api/reviews/daily/{date}:
    get:
      tags: [Reviews]
      summary: Get daily review by date
      parameters:
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Daily review for the specified date
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  date: { type: string, format: date }
                  note: { type: string }
                  completed_count: { type: integer }
        '404':
          description: No review found for this date

  # ─── Filters ───
...
  /api/rules:
    get:
      tags: [Rules]
      summary: List automation rules
      responses:
        '200':
          description: Array of rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Rule' }
    post:
      tags: [Rules]
      summary: Create an automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, trigger, action]
              properties:
                name: { type: string }
                trigger: { type: string }
                action: { type: string }
                config: { type: object }
                enabled: { type: integer, enum: [0, 1] }
      responses:
        '201':
          description: Created rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rule' }
  /api/rules/{id}:
    put:
      tags: [Rules]
      summary: Update a rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rule' }
    delete:
      tags: [Rules]
      summary: Delete a rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }

  # ─── Planner ───
...
components:
...
  schemas:
    Task:
      type: object
      properties:
        id: { type: integer }
        goal_id: { type: integer }
        title: { type: string }
        note: { type: string, nullable: true }
        status: { type: string, enum: [todo, doing, done] }
        priority: { type: integer, minimum: 0, maximum: 3 }
        position: { type: integer }
        due_date: { type: string, format: date, nullable: true }
        due_time: { type: string, nullable: true }
        recurring: { type: string, nullable: true }
        assigned_to: { type: string, nullable: true }
        my_day: { type: integer, enum: [0, 1] }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        estimated_minutes: { type: number, nullable: true }
        actual_minutes: { type: number, nullable: true }
        time_block_start: { type: string, nullable: true }
        time_block_end: { type: string, nullable: true }
        list_id: { type: integer, nullable: true }
        reminder_at: { type: string, nullable: true }
        tags:
          type: array
          items: { $ref: '#/components/schemas/Tag' }
        subtasks:
          type: array
          items: { $ref: '#/components/schemas/Subtask' }
        blocked_by:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              title: { type: string }
              status: { type: string }
    Subtask:
      type: object
      properties:
        id: { type: integer }
        task_id: { type: integer }
        title: { type: string }
        done: { type: integer, enum: [0, 1] }
        position: { type: integer }
    Tag:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: string, nullable: true }
...
    Note:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        pinned: { type: integer, enum: [0, 1] }
        color: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    InboxItem:
      type: object
      properties:
        id: { type: integer }
        text: { type: string }
        created_at: { type: string, format: date-time }
        processed: { type: integer, enum: [0, 1] }
    Review:
      type: object
      properties:
        id: { type: integer }
        week_start: { type: string, format: date }
        wins: { type: string }
        challenges: { type: string }
        next_week: { type: string }
        rating: { type: integer }
        created_at: { type: string, format: date-time }
...
    Rule:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        trigger: { type: string }
        action: { type: string }
        config: { type: object }
        enabled: { type: integer, enum: [0, 1] }
        created_at: { type: string, format: date-time }